:truck: rename: Move some methods from TokenProvider to TokenService #14
