+++ src/main/java/com/habitpay/habitpay/global/config/jwt/TokenProvider.java
@@ -1,21 +1,14 @@
 package com.habitpay.habitpay.global.config.jwt;
 
 import com.habitpay.habitpay.domain.member.domain.Member;
-import io.jsonwebtoken.Claims;
 import io.jsonwebtoken.Header;
 import io.jsonwebtoken.Jwts;
 import io.jsonwebtoken.SignatureAlgorithm;
 import lombok.RequiredArgsConstructor;
-import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
-import org.springframework.security.core.Authentication;
-import org.springframework.security.core.authority.SimpleGrantedAuthority;
-import org.springframework.security.core.userdetails.User;
 import org.springframework.stereotype.Service;
 
 import java.time.Duration;
-import java.util.Collections;
 import java.util.Date;
-import java.util.Set;
 
 @RequiredArgsConstructor
 @Service
@@ -51,34 +44,5 @@ public class TokenProvider {
         } catch (Exception e) {
             return false;
         }
-    }
-
-    public Authentication getAuthentication(String token) {
-        Claims claims = getClaims(token);
-        Set<SimpleGrantedAuthority> authorities = Collections.singleton(
-                new SimpleGrantedAuthority("ROLE_GUEST")
-        );
-
-        return new UsernamePasswordAuthenticationToken(
-                new User(claims.getSubject(), "", authorities),
-                token,
-                authorities);
-    }
-
-    public String getEmail(String token) {
-        Claims claims = getClaims(token);
-        return claims.getSubject();
-    }
-
-    public boolean getIsActive(String token) {
-        Claims claims = getClaims(token);
-        return claims.get("isActive", Boolean.class);
-    }
-
-    private Claims getClaims(String token) {
-        return Jwts.parser()
-                .setSigningKey(jwtProperties.getSecret())
-                .parseClaimsJws(token)
-                .getBody();
     }
 }
