+++ src/main/java/com/habitpay/habitpay/global/config/jwt/TokenService.java
@@ -2,15 +2,24 @@ package com.habitpay.habitpay.global.config.jwt;
 
 import com.habitpay.habitpay.domain.member.dao.MemberRepository;
 import com.habitpay.habitpay.domain.member.domain.Member;
+import io.jsonwebtoken.Claims;
+import io.jsonwebtoken.Jwts;
 import lombok.RequiredArgsConstructor;
+import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.authority.SimpleGrantedAuthority;
+import org.springframework.security.core.userdetails.User;
 import org.springframework.stereotype.Service;
 
 import java.time.Duration;
+import java.util.Collections;
 import java.util.Optional;
+import java.util.Set;
 
 @RequiredArgsConstructor
 @Service
 public class TokenService {
+    private final JwtProperties jwtProperties;
     private final TokenProvider tokenProvider;
     private final MemberRepository memberRepository;
 
@@ -21,5 +30,34 @@ public class TokenService {
         Member member = optionalMember.get();
 
         return tokenProvider.generateToken(member, Duration.ofHours(2));
+    }
+
+    public Authentication getAuthentication(String token) {
+        Claims claims = getClaims(token);
+        Set<SimpleGrantedAuthority> authorities = Collections.singleton(
+                new SimpleGrantedAuthority("ROLE_GUEST")
+        );
+
+        return new UsernamePasswordAuthenticationToken(
+                new User(claims.getSubject(), "", authorities),
+                token,
+                authorities);
+    }
+
+    private Claims getClaims(String token) {
+        return Jwts.parser()
+                .setSigningKey(jwtProperties.getSecret())
+                .parseClaimsJws(token)
+                .getBody();
+    }
+
+    public String getEmail(String token) {
+        Claims claims = getClaims(token);
+        return claims.getSubject();
+    }
+
+    public boolean getIsActive(String token) {
+        Claims claims = getClaims(token);
+        return claims.get("isActive", Boolean.class);
     }
 }
